chore: cleaned up SparseMatrix class and ran unit test successfully

+++ tests/unit/SparseMatrix_Test.h
@@ -45,7 +45,7 @@ namespace csmp
         Point<3> Point3U;
 
         std::vector<double> x, y, sol_y, solB, solD;
-        std::vector<uint32_t> sizetVector;
+        std::vector<size_t> sizetVector;
 
         double   fTolerance;
         const bool verbose_;
